refactor: extract common stepper logic out
[#OCD-3751]
